+++ lib/dry/cli.rb
@@ -18,6 +18,7 @@ module Dry
     require "dry/cli/spell_checker"
     require "dry/cli/banner"
     require "dry/cli/inflector"
+    require "dry/cli/dispatch"
 
     # Check if command
     #
@@ -128,7 +129,7 @@ module Dry
     # @api private
     def perform_command(arguments)
       command, args = parse(kommand, arguments, [])
-      command.call(**args)
+      command.call(**Dispatch.command_args_for(command, args))
     end
 
     # Invoke the CLI if registry passed
@@ -146,7 +147,7 @@ module Dry
       return stderr.puts(Usage.call(result)) unless command.respond_to?(:call)
 
       result.before_callbacks.run(command, **args)
-      command.call(**args)
+      command.call(**Dispatch.command_args_for(command, args))
       result.after_callbacks.run(command, **args)
     end