+++ spec/support/shared_examples/rendering.rb
@@ -15,6 +15,7 @@ RSpec.shared_examples "Rendering" do |cli|
         #{cmd} db [SUBCOMMAND]
         #{cmd} destroy [SUBCOMMAND]
         #{cmd} exec TASK [DIRS]                                        # Execute a task
+        #{cmd} externally-extended NAME                                # Command extended by third parties
         #{cmd} generate [SUBCOMMAND]
         #{cmd} greeting [RESPONSE]
         #{cmd} hello                                                   # Print a greeting
@@ -77,6 +78,7 @@ RSpec.shared_examples "Rendering" do |cli|
         #{cmd} db [SUBCOMMAND]
         #{cmd} destroy [SUBCOMMAND]
         #{cmd} exec TASK [DIRS]                                        # Execute a task
+        #{cmd} externally-extended NAME                                # Command extended by third parties
         #{cmd} generate [SUBCOMMAND]
         #{cmd} greeting [RESPONSE]
         #{cmd} hello                                                   # Print a greeting
@@ -107,6 +109,7 @@ RSpec.shared_examples "Rendering" do |cli|
         #{cmd} db [SUBCOMMAND]
         #{cmd} destroy [SUBCOMMAND]
         #{cmd} exec TASK [DIRS]                                        # Execute a task
+        #{cmd} externally-extended NAME                                # Command extended by third parties
         #{cmd} generate [SUBCOMMAND]
         #{cmd} greeting [RESPONSE]
         #{cmd} hello                                                   # Print a greeting